requirements.txt
#image preprocessing and video capturing
opencv-python
#for model training and loading the model
ultralytics
# for creating a text to speech engine
pyttsx3
# python
IPython
# Web framework
Flask==2.3.3
# For handling file uploads
Werkzeug==2.3.7
# For image processing
Pillow==10.0.1
# For JSON responses
flask-cors==4.0.0
# For base64 encoding
base64
# For datetime
datetime